Fixed byte array for conversion and precision.

+++ Decoder/Architecture/Domain Layer/Sensor/Measurements/Parameters/Soil Moisture/Permittivity.cs
@@ -1,21 +1,21 @@
 ﻿using System;
 namespace Decoder.Architecture.DomainLayer.Sensor.Measurements.Parameters.SoilMoisture
 {
-    public class Permittivity : IParameter<decimal>
+    public class Permittivity : IParameter<double>
     {
-        private readonly decimal measurement;
+        private readonly double measurement;
 
         #region Constructor:
 
         public Permittivity(double measurement) =>
-            this.measurement = (decimal)Math.Pow(
+            this.measurement = Math.Pow(
                 0.000000002887 * Math.Pow(
                     measurement / 10, 3) - 0.0000208 * Math.Pow(
                         measurement / 10, 2) + 0.05276 * (measurement / 10) - 43.39, 2);
 
         #endregion
 
-        public decimal Measurement => measurement;
+        public double Measurement => measurement;
         public string Units => null;
     }
 }
