+++ Decoder/Architecture/Domain Layer/Sensor/Measurements/Parameters/Soil Moisture/Temperature.cs
@@ -1,22 +1,22 @@
 ﻿using System;
 namespace Decoder.Architecture.DomainLayer.Sensor.Measurements.Parameters.SoilMoisture
 {
-    public class Temperature : IParameter<decimal>
+    public class Temperature : IParameter<double>
     {
-        private readonly decimal measurement;
+        private readonly double measurement;
         private readonly string units;
 
         #region Constructor:
 
         public Temperature(double measurement, string units)
         {
-            this.measurement = (decimal)((measurement - 32768) / 10);
+            this.measurement = ((measurement - 32768) / 10);
             this.units = units;
         }
 
         #endregion
 
-        public decimal Measurement => measurement;
+        public double Measurement => measurement;
 
         public string Units => units;
     }
